+++ user_extensions/apps.py
@@ -28,9 +28,15 @@ class UserExtensionsConfig(AppConfig):
                 SettingsType.COMMON: {
                     PluginSettings.RELATIVE_PATH: 'settings.common',
                 },
+                SettingsType.DEVSTACK: {
+                    PluginSettings.RELATIVE_PATH: 'settings.common',
+                },
             },
             ProjectType.LMS: {
                 SettingsType.COMMON: {
+                    PluginSettings.RELATIVE_PATH: 'settings.common',
+                },
+                SettingsType.DEVSTACK: {
                     PluginSettings.RELATIVE_PATH: 'settings.common',
                 },
             },